+++ src/allotropy/parsers/beckman_vi_cell_xr/vi_cell_xr_parser.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
 from enum import Enum
+import tempfile
 from typing import Any, Optional
+import zipfile
 
 import pandas as pd
 
@@ -40,6 +42,7 @@ from allotropy.parsers.beckman_vi_cell_xr.constants import (
 from allotropy.parsers.beckman_vi_cell_xr.vi_cell_xr_reader import ViCellXRReader
 from allotropy.parsers.utils.uuids import random_uuid_str
 from allotropy.parsers.vendor_parser import VendorParser
+from allotropy.types import IOType
 
 
 class SampleProperty(Enum):
@@ -62,9 +65,36 @@ def get_property_from_sample(
     return sample_property.data_type(value=value) if value else None
 
 
+def remove_style_xml_file(contents: IOType) -> IOType:
+    # Removes styles.xml from an xlsx file IO stream. xlsx files produced by VI-Cell XR
+    # instrument may have an invalid <fill> tag in their styles.xml file which causes a
+    # bug when reading with pandas (via openpyxl library).
+
+    # zipfile only accepts a filename, so write contents to a named temp file.
+    tmp = tempfile.NamedTemporaryFile()
+    file_contents = contents.read()
+    if isinstance(file_contents, str):
+        file_contents = file_contents.encode()
+    tmp.write(file_contents)
+
+    # Write zip contents to a new file, skipping styles.xml
+    new = tempfile.NamedTemporaryFile()
+    with zipfile.ZipFile(tmp.name) as zin:
+        with zipfile.ZipFile(new.name, "w") as zout:
+            for item in zin.infolist():
+                if item.filename == "xl/styles.xml":
+                    continue
+                zout.writestr(item, zin.read(item.filename))
+
+    return open(new.name, "rb")
+
+
 class ViCellXRParser(VendorParser):
     def to_allotrope(self, named_file_contents: NamedFileContents) -> Model:
         contents, filename = named_file_contents
+
+        if filename.endswith("xlsx"):
+            contents = remove_style_xml_file(contents)
         reader = ViCellXRReader(contents)
 
         return Model(