+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add try_float_or_nan util and fix bug with evaluating 0 as NaN
 ### Fixed
 - Cast sample identifier to string when saving it in SoftmaxPro parser
+- Handle style bug in xlsx files produced by VI-Cell XR instrument
 ### Changed
 ### Deprecated
 ### Removed